+++ parser_library/test/context/system_variable_test.cpp
@@ -263,34 +263,6 @@ TEST(system_variable, sysstmt_copy)
     EXPECT_EQ(get_var_value<C_t>(a.hlasm_ctx(), "VAR2"), "00000019");
 }
 
-TEST(system_variable, sysstmt_ainsert)
-{
-    std::string input = R"(
-    MACRO
-    MAC_AIN
-    AINSERT '       MACRO',BACK
-    AINSERT '       MAC',BACK
-    AINSERT '       GBLC &&A',BACK
-    AINSERT '&&A    SETC ''&&SYSSTMT''',BACK
-    AINSERT '       MEND',BACK
-    AINSERT '&&B    SETC ''&&SYSSTMT''',BACK
-    MEND
-    
-    GBLC &A
-    MAC_AIN
-    MAC
-    END
-)";
-
-    analyzer a(input);
-    a.analyze();
-    a.collect_diags();
-    EXPECT_EQ(a.diags().size(), (size_t)0);
-
-    EXPECT_EQ(get_var_value<C_t>(a.hlasm_ctx(), "A"), "00000029");
-    EXPECT_EQ(get_var_value<C_t>(a.hlasm_ctx(), "B"), "00000027");
-}
-
 TEST(system_variable, sysstmt_aread)
 {
     std::string input = R"(
@@ -312,32 +284,3 @@ TEST(system_variable, sysstmt_aread)
 
     EXPECT_EQ(get_var_value<A_t>(a.hlasm_ctx(), "A"), 13);
 }
-
-// TODO uncomment when AINSERT grammer with apostrophes is fixed; Consider moving this to AINSERT test group
-// TEST(system_variable, sysstmt_ainsert_02)
-//{
-//    std::string input = R"(
-//    MACRO
-//    MAC_AIN
-//    AINSERT '       MACRO',BACK
-//    AINSERT '       MAC',BACK
-//    AINSERT '       GBLC &&A',BACK
-//    AINSERT '&&A    SETC ''&SYSSTMT''',BACK
-//    AINSERT '       MEND',BACK
-//    AINSERT '&&B    SETC ''&SYSSTMT''',BACK
-//    MEND
-//
-//    GBLC &A
-//    MAC_AIN
-//    MAC
-//    END
-//)";
-//
-//    analyzer a(input);
-//    a.analyze();
-//    a.collect_diags();
-//    EXPECT_EQ(a.diags().size(), (size_t)0);
-//
-//    EXPECT_EQ(get_var_value<C_t>(a.hlasm_ctx(), "A"), "00000017");
-//    EXPECT_EQ(get_var_value<C_t>(a.hlasm_ctx(), "B"), "00000019");
-//}