+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -11,6 +11,7 @@
 - Behavior of currently supported subscripted system variables corrected
 - Cross-section relative immediate references should be only warned upon
 - Incorrect module layout computed when ORG instruction is used in sections with multiple location counters
+- AINSERT can now do immediate variable evaluatation correctly
 - Highlighting of single character strings (which could represent data attributes - e.g. `I'`, `L'`, `S'` and others) is fixed
 
 ## [1.0.0](https://github.com/eclipse/che-che4z-lsp-for-hlasm/compare/0.15.1...1.0.0) (2022-01-31)